Move event_absolute_position to odra-utils crate
