+++ utils/Cargo.toml
@@ -5,3 +5,4 @@ edition = "2021"
 
 [dependencies]
 convert_case = "0.5.0"
+odra-types = { path = "../types" }
